requirements.txt
beautifulsoup4==4.14.3
fastapi==0.135.1
httpx==0.28.1
numpy==2.4.3
pandas==3.0.1
pydantic==2.12.5
torch==2.6.0
python-dotenv==1.0.1
openai>=1.0.0,<2.0.0
transformers==5.3.0
uvicorn==0.42.0

pymysql~=1.1.2